test: move all db tests to dbtest package
prerequisite before moving to a separate go module. Tests, which were
not executed in a real db were kept in the same place as before
