+++ dbtests/postgresql_test.go
@@ -1,11 +1,12 @@
 //go:build postgresql
 
-package testfixtures
+package dbtests
 
 import (
 	"os"
 	"testing"
 
+	"github.com/go-testfixtures/testfixtures/v3"
 	_ "github.com/jackc/pgx/v4/stdlib"
 	_ "github.com/lib/pq"
 )
@@ -15,14 +16,14 @@ func TestPostgreSQL(t *testing.T) {
 }
 
 func TestPostgreSQLWithAlterConstraint(t *testing.T) {
-	testPostgreSQL(t, UseAlterConstraint())
+	testPostgreSQL(t, testfixtures.UseAlterConstraint())
 }
 
 func TestPostgreSQLWithDropConstraint(t *testing.T) {
-	testPostgreSQL(t, UseDropConstraint())
+	testPostgreSQL(t, testfixtures.UseDropConstraint())
 }
 
-func testPostgreSQL(t *testing.T, additionalOptions ...func(*Loader) error) {
+func testPostgreSQL(t *testing.T, additionalOptions ...func(*testfixtures.Loader) error) {
 	t.Helper()
 	for _, dialect := range []string{"postgres", "pgx"} {
 		db := openDB(t, dialect, os.Getenv("PG_CONN_STRING"))
