+++ dbtests/testdata/schema/cockroachdb.sql
