+++ dbtests/testdata/schema/spanner.sql
