+++ dbtests/testdata/schema/sqlserver.sql
